Fixed the buggy collisions and resolved the warnings.

+++ AFTERMATH -- SFML/AFTERMATH -- SFML/EditorState.cpp
@@ -13,8 +13,8 @@ void EditorState::initVariables()
 void EditorState::initView()
 {
 	this->view.setSize(sf::Vector2f(
-		this->stateData->gfxSettings->resolution.width, 
-		this->stateData->gfxSettings->resolution.height));
+		static_cast<float>(this->stateData->gfxSettings->resolution.width),
+			static_cast<float>(this->stateData->gfxSettings->resolution.height)));
 	this->view.setCenter(
 		this->stateData->gfxSettings->resolution.width / 2.f,
 		this->stateData->gfxSettings->resolution.height / 2.f
